refactor(controller): remove id parsing and try/catch

+++ back/src/main/java/com/openclassrooms/starterjwt/services/UserService.java
@@ -1,8 +1,13 @@
 package com.openclassrooms.starterjwt.services;
 
+import com.openclassrooms.starterjwt.exception.UnauthorizedOperationException;
 import com.openclassrooms.starterjwt.models.User;
 import com.openclassrooms.starterjwt.repository.UserRepository;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Service;
+
+import java.util.Objects;
 
 @Service
 public class UserService {
@@ -13,6 +18,17 @@ public class UserService {
     }
 
     public void delete(Long id) {
+        User user = this.findById(id);
+
+        UserDetails userDetails = (UserDetails) SecurityContextHolder
+                .getContext()
+                .getAuthentication()
+                .getPrincipal();
+
+        if (Objects.equals(userDetails.getUsername(), user.getEmail())) {
+            throw new UnauthorizedOperationException("You cannot delete your own account");
+        }
+
         this.userRepository.deleteById(id);
     }
